Saman10/23

+++ app/api/review_routes.py
@@ -1,43 +1,23 @@
-from flask import Blueprint, jsonify, request
-from flask_login import login_required
+from flask import Blueprint, request, jsonify, render_template, redirect, url_for
+from flask_login import login_required, current_user
 from app.models import Review
 from app import db
+from app.forms.review_form import ReviewForm
+
 
 review_routes = Blueprint('reviews', __name__)
 
-# Retrieve reviews for a specific product
-@review_routes.route('/products/<int:productId>/reviews', methods=['GET'])
-def get_reviews(productId):
-    reviews = Review.query.filter_by(productId=productId).all()
-    return {'reviews': [review.to_dict() for review in reviews]}
 
-
-# Add a review for a specific product (for authenticated users)
-@review_routes.route('/products/<int:productId>/reviews', methods=['POST'])
-@login_required
-def add_review(productId):
-    data = request.get_json()
-
-    new_review = Review(
-        userId=data['userId'],
-        productId=productId,
-        description=data['description'],
-        rating=data['rating'],
-        dateCreated=db.func.now()
-    )
-
-    db.session.add(new_review)
-    db.session.commit()
-
-    return new_review.to_dict()
-
-# Update a specific review (for authenticated users)
-@review_routes.route('/reviews/<int:reviewId>', methods=['PUT'])
+# Update a specific review
+@review_routes.route('/<int:reviewId>', methods=['PUT'])
 @login_required
 def update_review(reviewId):
     review = Review.query.get(reviewId)
     if not review:
         return ('Review not found', 404)
+
+    if review.userId != current_user.id:
+        return ('Not authorized to update this review', 403)
 
     data = request.get_json()
 
@@ -48,15 +28,25 @@ def update_review(reviewId):
 
     return review.to_dict()
 
-# Delete a specific review (for authenticated users)
-@review_routes.route('/reviews/<int:reviewId>', methods=['DELETE'])
+# Delete a specific review
+@review_routes.route('/<int:reviewId>', methods=['DELETE'])
 @login_required
 def delete_review(reviewId):
     review = Review.query.get(reviewId)
     if not review:
         return ('Review not found', 404)
 
+    if review.userId != current_user.id:
+        return ('Not authorized to delete this review', 403)
+
     db.session.delete(review)
     db.session.commit()
 
     return {'message': 'Review deleted'}
+
+# Get all the review from a logged in user
+@review_routes.route('/user', methods=['GET'])
+@login_required
+def get_user_reviews():
+    user_reviews = Review.query.filter_by(userId=current_user.id).all()
+    return jsonify({'user_reviews': [review.to_dict() for review in user_reviews]})
